API/storages: move storage related Exceptions to 'storages' submodule.
