+++ Utils/API/server/lib/dkb/api/exceptions.py
@@ -147,24 +147,6 @@ class InvalidArgument(MethodException):
             super(InvalidArgument, self).__init__(method, reason)
 
 
-class StorageException(DkbApiException):
-    """ Base exception for storage failures. """
-    code = 580
-    details = "Failed to get data from DKB storage."
-
-
-class StorageClientException(StorageException):
-    """ Exception indicating failure when creating storage client. """
-    code = 581
-
-    def __init__(self, storage, reason=None):
-        message = "Failed to initialize storage client for '%s'." % storage
-        if reason:
-            message += " Reason: %s" % reason
-        self.details = message
-        super(StorageClientException, self).__init__(message)
-
-
 class ConfigurationError(DkbApiException):
     """ Base exception for server configuration errors. """
     code = 590
